Add next-auth for authentication
This commit adds next-auth for authentication. It also adds a Header
component that will be used to display the user's state and a logout
button.

The next-auth package is configured to use the credentials provider
and the JWT provider. The JWT provider is used to store the user's data
and api JWT token.